Update readme.md
Howdy, I've made some edits and suggestions. 

I also agree with the comment I saw in the thread about having a clear process to have a process defined to accept, integrate, and release updated code.

I'm also not quite sure how to word it, but do we want to maintain an "officially vetted" release along with a "latest beta" release so the Github community can help test?

I don't think it belongs in this document, but if an official product ends up using this location as the main distribution point, we should still make sure it has a coversheet so it can be listing on our Mining website.